reorganize the binary locations and add an empty raywa.rs
