+++ wasm/src/cmd/raywa.rs
@@ -11,19 +11,16 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use std::net::UdpSocket;
+use wasm_on_ray::config;
+use wasm_on_ray::ray;
+use wasm_on_ray::util;
 
-pub fn get_node_ip_address(address: &str) -> String {
-    // convert str to String
-    let mut ip_address = address.to_string();
-    if ip_address.is_empty() {
-        ip_address = "8.8.8.8:53".to_string();
-    }
-    // use udp resolver to get local ip address
-    let socket = UdpSocket::bind("0.0.0.0:0").expect("Failed to bind socket");
-    socket
-        .connect(ip_address)
-        .expect("Failed to connect to DNS server");
-    let local_addr = socket.local_addr().expect("Failed to get local address");
-    local_addr.ip().to_string()
+use anyhow::Result;
+use clap::Parser;
+use tracing_subscriber;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    tracing_subscriber::fmt().init();
+    Ok(())
 }
