+++ wasm/src/ray/core/global_state_accessor.rs
